[TASK] Move composer functionality to an abstract class
To implement more composer tasks, the functionality how to build a composer command
is now in an abstract class. The install task extends it and override command and arguments.
